Add xml_register. All mss object are now registered in a xml global object.
This permits to manage properly name of mss ressource and offers xpath
access to them.

for instance, to get the uri of all states:
mss.xml_register.XmlRegister.find_all_elts("//*[@ressource='state']")
Which returns a list of uri.

Moreover, this also bring xml export which permits to do some work in
the client.

It was hard to choose a xml representation. Choices were :
1) <lifecycle name=Mysql>
   <state name=Acitve/>
   </lifecycle>
OR
2) <Mysql ressource=lifecycle>
   <Active ressource=state/>
   </Mysql>

The second choice were choosen because it permits to generate clean uri such as
/Mysql/Active
/Http/Active
while the first choice generates
/lifecycle/state
/lifecycle[2]/state