[smart] validate values received by the generator
Variables aren't always validated from the scope varaible objects.

In the case of smartonic the scope variables are filled, serialized,
validated via provide_call_validate and finally the scope variables are updated
with the result.

With the xmpp master, the client send directly the variables values to
validate. So the values are validated via provide_call_validate and finally
scope variables are updated with the result.