[xmpp] Send log level info over xmpp
Agents are now sending the log level information with their log messages. The
log level information is described by the ArmonicLog stanza.

This stanza can be added to any Message stanza and is processed in
XMPPClientSync.handle_armonic_log

If the agent is in devel mode, the client will display the agent debug messages
only when on debug mode.