[smart] don't update the variable value from the json generated by agent.
Indeed, the agent doen't have to change this value. In fact, we don't
do this because the json variable specification is not sufficient to
manage multiplicity of variables. See commit message also.

To send variable to agent, the following format is used:
[(/xpath/of/variable1, {0: val1, 1: val2}),
 (/xpath/of/variable2, {0: val1, 1: val2})]

But when the agent returns the validated variables, the format is:
[(/xpath/of/variable1, val1),
 (/xpath/of/variable2, val1)
 (/xpath/of/variable1, val2),
 (/xpath/of/variable2, val2)
]
where we are not able to find out the require index (0 or 1 in this
example) when multiplicity is used.